Move function getNumberOfPages into separate module
