+++ src/app/shared/lib/github-paginator-parser.ts
@@ -1,3 +1,18 @@
+import { GithubResponse } from 'src/app/core/models/github/github-response.model';
+
+/**
+ * Get the number of paginated pages of issues specified in a GitHubResponse
+ * @param response
+ */
+export function getNumberOfPages<T>(response: GithubResponse<T>): number {
+  let numberOfPages = 1;
+  if (response.headers.link) {
+    const paginatedData = githubPaginatorParser(response.headers.link);
+    numberOfPages = +paginatedData['last'] || 1;
+  }
+  return numberOfPages;
+}
+
 /**
  * Will return in the format of { paginateAction: number }
  * Example { next: '15', last: '34', first: '1', prev: '13' }
@@ -7,7 +22,7 @@
  * @param linkStr represent the pagination string provided by github API.
  *
  */
-export function githubPaginatorParser(linkStr: string) {
+function githubPaginatorParser(linkStr: string) {
   return linkStr.split(',').map((paginateItem) => {
     return paginateItem.split(';').map((curr, idx) => {
       if (idx === 0) {
